Set open file limits in line with Nginx best practice
Prior to this change, we would run out of file descriptors because Nginx was configured for 4 workers with 1024 files each, but the default number of open files allowed by the systemd unit file (not set) is 1024, so we could run out of file descriptors when under a DDoS.

This change sets a number of values:
 * worker_processes auto - this allows nginx to use a process per CPU
 * worker_rlimit_nofile 1024 - limit each nginx process to 1024 FDs
 * LimitNOFILE=32768 - systemd unit file sets a larger FD limit

This gives us enough open files for up to 32 processes. Since our worker machines are much smaller than this, we should never run out of file descriptors.